Convert GradleVersionUtil
